Refactored code to extract bonus product item sub-component
